memorify_backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.0rc1
Django==4.1.3
django-cors-headers==4.2.0
django-extensions==3.2.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
frozendict==2.3.8
future==0.18.3
html5lib==1.1
idna==2.10
inflection==0.5.1
lxml==4.9.3
multitasking==0.0.11
numpy==1.25.0
oauthlib==3.1.0
pandas==2.0.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.20
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.31.0
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.15.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.2
views==0.3
webencodings==0.5.1
yfinance==0.2.22
selenium==4.17.2